+++ packages/broker/test/utils.ts
@@ -255,7 +255,9 @@ export async function startStorageNode(
         },
     })
     try {
-        await client.createOrUpdateNodeInStorageNodeRegistry(`http://127.0.0.1:${httpPort}`)
+        await client.setStorageNodeMetadata({
+            http: `http://127.0.0.1:${httpPort}`
+        })
         await createAssignmentStream(client)
     } finally {
         client?.destroy()