+++ packages/client/test/test-utils/fake/FakeStorageNodeRegistry.ts
@@ -4,7 +4,7 @@ import { StreamIDBuilder } from '../../../src/StreamIDBuilder'
 import { DOCKER_DEV_STORAGE_NODE } from '../../../src/ConfigTest'
 import { FakeStorageNode } from './FakeStorageNode'
 import { ActiveNodes } from './ActiveNodes'
-import { StorageNodeRegistry } from '../../../src/StorageNodeRegistry'
+import { StorageNodeMetadata, StorageNodeRegistry } from '../../../src/StorageNodeRegistry'
 import { Stream } from '../../../src/Stream'
 import { Multimap } from '../utils'
 import { StreamRegistry } from '../../../src/StreamRegistry'
@@ -88,17 +88,7 @@ export class FakeStorageNodeRegistry implements Omit<StorageNodeRegistry,
     }
 
     // eslint-disable-next-line class-methods-use-this
-    createOrUpdateNodeInStorageNodeRegistry(_httpUrl: string): Promise<void> {
-        throw new Error('not implemented')
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    removeNodeFromStorageNodeRegistry(): Promise<void> {
-        throw new Error('not implemented')
-    }
-
-    // eslint-disable-next-line class-methods-use-this
-    removeStreamFromStorageNode(_streamIdOrPath: string, _nodeAddress: string): Promise<void> {
+    async setStorageNodeMetadata(_metadata: StorageNodeMetadata | undefined): Promise<void> {
         throw new Error('not implemented')
     }